requirements.txt
# General AI Trading System - Dependencies
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
jinja2==3.1.5
python-multipart==0.0.18

# Data
ccxt==4.4.54
yfinance==0.2.51
pandas==2.2.3
numpy==1.26.4

# Technical Analysis
pandas-ta==0.3.14b1

# Machine Learning
scikit-learn==1.6.1
xgboost==2.1.3
torch==2.5.1
joblib==1.4.2

# Utilities
python-dotenv==1.0.1
aiofiles==24.1.0

# Databricks
databricks-sql-connector==4.0.2
mlflow==2.21.3
databricks-sdk==0.49.0
